app_back/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
requests>=2.31.0
rapidfuzz>=3.6.0
